requirements.txt
Flask
google-cloud-firestore
google-cloud-tasks
google-api-core
solapi